+++ packages/site/src/content/BottomSheet/index.tsx
@@ -1,6 +1,8 @@
 import Content from "@atlantis/docs/components/BottomSheet/BottomSheet.stories.mdx";
 import MobileProps from "./BottomSheet.props-mobile.json";
+import Notes from "./BottomSheetNotes.mdx";
 import { ContentExport } from "../../types/content";
+import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
@@ -48,7 +50,10 @@ alert('will show bottom sheet in mobile');
   links: [
     {
       label: "Storybook",
-      url: "http://localhost:6006/?path=/docs/components-utilities-BottomSheet-web--docs",
+      url: getStorybookUrl(
+        `?path=/docs/components-selections-bottomsheet--docs`,
+      ),
     },
   ],
+  notes: () => <Notes />,
 } as const satisfies ContentExport;