+++ packages/site/src/content/FeatureSwitch/index.tsx
@@ -1,6 +1,7 @@
 import Content from "@atlantis/docs/components/FeatureSwitch/FeatureSwitch.stories.mdx";
 import Props from "./FeatureSwitch.props.json";
 import { ContentExport } from "../../types/content";
+import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
@@ -32,7 +33,9 @@ export default {
   links: [
     {
       label: "Storybook",
-      url: "http://localhost:6006/?path=/docs/components-utilities-FeatureSwitch-web--docs",
+      url: getStorybookUrl(
+        `?path=/docs/components-selections-featureswitch--docs`,
+      ),
     },
   ],
 } as const satisfies ContentExport;