+++ packages/site/src/content/InputFile/index.tsx
@@ -1,6 +1,8 @@
 import Content from "@atlantis/docs/components/InputFile/InputFile.stories.mdx";
 import Props from "./InputFile.props.json";
+import Notes from "./InputFileNotes.mdx";
 import { ContentExport } from "../../types/content";
+import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
@@ -13,7 +15,10 @@ export default {
   links: [
     {
       label: "Storybook",
-      url: "http://localhost:6006/?path=/docs/components-utilities-InputFile-web--docs",
+      url: getStorybookUrl(
+        `?path=/docs/components-forms-and-inputs-inputfile--docs`,
+      ),
     },
   ],
+  notes: () => <Notes />,
 } as const satisfies ContentExport;