+++ packages/site/src/content/InputNumber/index.tsx
@@ -1,7 +1,9 @@
 import Content from "@atlantis/docs/components/InputNumber/InputNumber.stories.mdx";
 import Props from "./InputNumber.props.json";
 import MobileProps from "./InputNumber.props-mobile.json";
+import Notes from "./InputNumberNotes.mdx";
 import { ContentExport } from "../../types/content";
+import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
@@ -21,7 +23,10 @@ export default {
   links: [
     {
       label: "Storybook",
-      url: "http://localhost:6006/?path=/docs/components-utilities-InputNumber-web--docs",
+      url: getStorybookUrl(
+        `?path=/docs/components-forms-and-inputs-inputnumber--docs`,
+      ),
     },
   ],
+  notes: () => <Notes />,
 } as const satisfies ContentExport;