+++ packages/site/src/content/InputPhoneNumber/index.tsx
@@ -1,6 +1,7 @@
 import Content from "@atlantis/docs/components/InputPhoneNumber/InputPhoneNumber.stories.mdx";
 import Props from "./InputPhoneNumber.props.json";
 import { ContentExport } from "../../types/content";
+import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
@@ -19,7 +20,9 @@ export default {
   links: [
     {
       label: "Storybook",
-      url: "http://localhost:6006/?path=/docs/components-utilities-InputPhoneNumber-web--docs",
+      url: getStorybookUrl(
+        `?path=/docs/components-forms-and-inputs-inputphonenumber--docs`,
+      ),
     },
   ],
 } as const satisfies ContentExport;