+++ packages/site/src/content/ProgressBar/index.tsx
@@ -2,6 +2,7 @@ import ProgressBarContent from "@atlantis/docs/components/ProgressBar/ProgressBa
 import Props from "./ProgressBar.props.json";
 import MobileProps from "./ProgressBar.props-mobile.json";
 import { ContentExport } from "../../types/content";
+import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <ProgressBarContent />,
@@ -9,13 +10,15 @@ export default {
   mobileProps: MobileProps,
   component: {
     element: `<ProgressBar totalSteps={100} currentStep={66} />`,
-    mobileElement: `<ProgressBar total={5} current={1} inProgress={2} />`,
+    mobileElement: `<View style={{width:'100%'}}><ProgressBar total={5} current={1} inProgress={2} /></View>`,
   },
   title: "ProgressBar",
   links: [
     {
       label: "Storybook",
-      url: "http://localhost:6006/?path=/docs/components-utilities-ProgressBar-web--docs",
+      url: getStorybookUrl(
+        `?path=/docs/components-status-and-feedback-progressbar--docs`,
+      ),
     },
   ],
 } as const satisfies ContentExport;